requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.5
djangorestframework==3.11.0
jwt==1.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycparser==2.20
python-decouple==3.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1